bugfix for license unlink

+++ app/grails-app/services/de/laser/SubscriptionService.groovy
@@ -1578,12 +1578,15 @@ class SubscriptionService {
             }
         }
         else if(unlink && curLink) {
-            License lic = genericOIDService.resolveOID(curLink.source)
+            String sourceOID = curLink.source
+            License lic = genericOIDService.resolveOID(sourceOID)
             curLink.delete() //delete() is void, no way to check whether errors occurred or not
-            Set<Links> linkedSubs = Links.executeQuery("select li from Links li where li.linkType = :linkType and li.destination in (select concat('${Subscription.class.name}:',oo.sub.id) from OrgRole oo where oo.roleType in (:subscrTypes) and oo.org = :subscr)",[linkType:RDStore.LINKTYPE_LICENSE,subscrTypes:[RDStore.OR_SUBSCRIBER_CONS,RDStore.OR_SUBSCRIBER_CONS_HIDDEN],subscr:subscr])
-            if(linkedSubs.size() < 1) {
-                log.info("no more license <-> subscription links between org -> removing licensee role")
-                OrgRole.executeUpdate("delete from OrgRole oo where oo.lic = :lic and oo.org = :subscriber",[lic:lic,subscriber:subscr])
+            if(sub.getCalculatedType() == CalculatedType.TYPE_PARTICIPATION) {
+                Set<Links> linkedSubs = Links.executeQuery("select li from Links li where li.source = :lic and li.linkType = :linkType and li.destination in (select concat('${Subscription.class.name}:',oo.sub.id) from OrgRole oo where oo.roleType in (:subscrTypes) and oo.org = :subscr)", [lic: sourceOID, linkType: RDStore.LINKTYPE_LICENSE, subscrTypes: [RDStore.OR_SUBSCRIBER_CONS, RDStore.OR_SUBSCRIBER_CONS_HIDDEN], subscr: subscr])
+                if (linkedSubs.size() < 1) {
+                    log.info("no more license <-> subscription links between org -> removing licensee role")
+                    OrgRole.executeUpdate("delete from OrgRole oo where oo.lic = :lic and oo.org = :subscriber", [lic: lic, subscriber: subscr])
+                }
             }
             success = true
         }
